+++ src/main/java/com/example/mrrag/app/controller/GraphApiController.java
@@ -3,7 +3,6 @@ package com.example.mrrag.app.controller;
 import com.example.mrrag.app.controller.requestDTO.RemoteProjectRequest;
 import com.example.mrrag.app.source.GitLabLocalSourceProvider;
 import com.example.mrrag.graph.GraphBuildStats;
-import com.example.mrrag.app.service.AstGraphService;
 import com.example.mrrag.graph.model.EdgeKind;
 import com.example.mrrag.graph.model.NodeKind;
 import com.example.mrrag.graph.model.ProjectGraph;
