+++ src/main/java/com/example/mrrag/graph/cache/IncrementalGraphBuilder.java
@@ -2,7 +2,6 @@ package com.example.mrrag.graph.cache;
 
 import com.example.mrrag.app.source.ProjectSource;
 import com.example.mrrag.app.source.ProjectSourceProvider;
-import com.example.mrrag.graph.GraphBuilder;
 import com.example.mrrag.graph.model.ProjectGraph;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
