+++ src/main/java/com/example/mrrag/review/ReviewService.java
@@ -2,7 +2,6 @@ package com.example.mrrag.review;
 
 import com.example.mrrag.app.controller.requestDTO.RemoteProjectRequest;
 import com.example.mrrag.app.controller.requestDTO.ReviewRequest;
-import com.example.mrrag.app.service.AstGraphService;
 import com.example.mrrag.app.source.GitLabLocalSourceProvider;
 import com.example.mrrag.graph.model.ProjectGraph;
 import com.example.mrrag.review.model.*;
